refactor: add scan module with unified diagnostics
Introduce a shared scan module owning ScanDiagnostics/ScanFailure and the
scan thread pool. Alias the per-feature diagnostics types to the shared
ones so usage and analysis stop duplicating the collection-failure shape.
