+++ src/lib.rs
@@ -35,6 +35,7 @@ pub mod logging;
 pub mod models;
 pub mod pricing;
 pub mod quota;
+pub mod scan;
 pub mod session;
 pub mod summary_cache;
 pub mod update;
